+++ src/features/community/components/new-moment-form/index.ts
@@ -1,0 +1,1 @@
+export * from './NewMomentForm';
